+++ .atomist/handlers/command/ExportCommonHandlers.ts
@@ -1,0 +1,4 @@
+import * as CommonHandlers from "@atomist/rugs/operations/CommonHandlers";
+
+export const boo = new CommonHandlers.GenericErrorHandler();
+export const yay = new CommonHandlers.GenericSuccessHandler();